changed doubles to floats

+++ src/common.h
@@ -14,19 +14,19 @@ using namespace glm;
 const bool RADOMIZE_OFFSET = false;
 const int SOLVER_ITERS = 4;
 
-const dvec3 ACCEL_GRAVITY(0, -9.8, 0);
-const vector<dvec3> EXTERNAL_ACCELRATIONS = {ACCEL_GRAVITY};
+const vec3 ACCEL_GRAVITY(0, -9.8, 0);
+const vector<vec3> EXTERNAL_ACCELRATIONS = {ACCEL_GRAVITY};
 
-const double PARTICLE_RADIUS = 0.05;
-const double PARTICLE_MASS = 1;
-const double REST_DENSITY = 6378;
-const double VISCOSITY = 0.01;
+const float PARTICLE_RADIUS = 0.05;
+const float PARTICLE_MASS = 1;
+const float REST_DENSITY = 6378;
+const float VISCOSITY = 0.01;
 
-const double KERNEL_RADIUS = 0.1;
-const double EPS = 600;
-const double PRESSURE_STRENGTH = 0.1;
-const double PRESSURE_POW = 4;
-const dvec3 DELTA_Q(0.2 * KERNEL_RADIUS, 0, 0);
+const float KERNEL_RADIUS = 0.1;
+const float EPS = 600;
+const float PRESSURE_STRENGTH = 0.1;
+const float PRESSURE_POW = 4;
+const vec3 DELTA_Q(0.2 * KERNEL_RADIUS, 0, 0);
 
 const float EPS_F = 0.00001f;
